proyectos_alternativos/requirements.txt
# PyPI-only requirements for proyectos_alternativos (FastAPI backends)
# Generated by scanning your uploaded ZIP project code (excluding embedded venv/site-packages).

fastapi
uvicorn[standard]
python-dotenv
httpx
